flask/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# config\settings.py: 3
celery == 4.4.0

# cli\cli.py: 1
# cli\commands\cmd_scripts.py: 1
click == 7.0

# quickslack\app.py: 5
flask == 1.1.1

# quickslack\extensions.py: 1
flask-debugtoolbar == 0.10.1

# quickslack\extensions.py: 2
flask-sqlalchemy == 2.4.1

# config\settings.py: 1
python-decouple == 3.3

# setup.py: 1
setuptools == 44.0.0

# not found by pigar
gunicorn
psycopg2-binary
redis